関数再配置おわり
